Per arch build and inlining of cpuid_x86.
